feat: synthese DS et Orga revues

+++ packages/frontend-usagers/src/components/personnes.vue
@@ -6,6 +6,7 @@
       <DsfrTable :headers="headersToDisplay" :rows="personnesToDisplay" />
     </div>
     <DsfrButton
+      v-if="props.modifiable"
       ref="modalOrigin"
       label="Ajouter un représentant légal"
       size="sm"
@@ -16,15 +17,18 @@
       ref="modal"
       name="test"
       :opened="modalPersonne.opened"
-      :title="props.title"
+      :title="props.titre"
       size="md"
       @close="onClose"
     >
       <Personne
+        :modifiable="props.modifiable"
         :personne="personne"
         :show-adresse="props.showAdresse"
         :show-telephone="props.showTelephone"
         :show-email="props.showEmail"
+        :show-button="props.modifiable"
+        :validate-on-mount="!props.modifiable"
         @valid="updatePersonne"
       ></Personne>
     </DsfrModal>
@@ -34,6 +38,7 @@
 <script setup>
 const props = defineProps({
   personnes: { type: Array, required: true },
+  modifiable: { type: Boolean, default: true },
   showAdresse: { type: Boolean, default: false, required: false },
   showTelephone: { type: Boolean, default: false, required: false },
   showEmail: { type: Boolean, default: false, required: false },
@@ -62,16 +67,18 @@ const personnesToDisplay = computed(() => {
     displayedFields.forEach((f) => {
       row.push(p[f]);
     });
-    row.push({
-      component: "DsfrButton",
-      class: "fr-icon-delete-fill",
-      tertiary: true,
-      noOutline: true,
-      onClick: (event) => {
-        event.stopPropagation();
-        deleteItem(index);
-      },
-    });
+    if (props.modifiable) {
+      row.push({
+        component: "DsfrButton",
+        class: "fr-icon-delete-fill",
+        tertiary: true,
+        noOutline: true,
+        onClick: (event) => {
+          event.stopPropagation();
+          deleteItem(index);
+        },
+      });
+    } else row.push({});
     return {
       rowData: row,
       rowAttrs: { class: "pointer", onClick: () => editItem(index) },
